+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/playlist/HlsPlaylistParser.java
@@ -714,11 +714,9 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
       } else if (line.equals(TAG_DISCONTINUITY)) {
         relativeDiscontinuitySequence++;
       } else if (line.startsWith(TAG_PROGRAM_DATE_TIME)) {
-        if (playlistStartTimeUs == 0) {
-          long programDatetimeUs =
-              C.msToUs(Util.parseXsDateTime(line.substring(line.indexOf(':') + 1)));
-          playlistStartTimeUs = programDatetimeUs - segmentStartTimeUs;
-        }
+        long programDatetimeUs =
+            C.msToUs(Util.parseXsDateTime(line.substring(line.indexOf(':') + 1)));
+        playlistStartTimeUs = programDatetimeUs - segmentStartTimeUs;
       } else if (line.equals(TAG_GAP)) {
         hasGapTag = true;
       } else if (line.equals(TAG_INDEPENDENT_SEGMENTS)) {